# Add caching for tab size hints