--- a/qutebrowser/mainwindow/tabwidget.py
+++ b/qutebrowser/mainwindow/tabwidget.py
@@ -447,13 +447,14 @@ class TabBar(QTabBar):
         Return:
             A QSize of the smallest tab size we can make.
         """
-        return self.__minimumTabSizeHintHelper(self.tabText(index),
-                                               self.tabIcon(index), ellipsis)
+        return self._minimum_tab_size_hint_helper(self.tabText(index),
+                                                  self.tabIcon(index),
+                                                  ellipsis)
 
     @functools.lru_cache(maxsize=100)
-    def __minimumTabSizeHintHelper(self, tab_text: str,
-                                   icon,
-                                   ellipsis: bool) -> QSize:
+    def _minimum_tab_size_hint_helper(self, tab_text: str,
+                                      icon,
+                                      ellipsis: bool) -> QSize:
         """Helper function to cache tab results."""
         text = '\u2026' if ellipsis else tab_text
         # Don't ever shorten if text is shorter than the ellipsis