--- a/qutebrowser/misc/utilcmds.py
+++ b/qutebrowser/misc/utilcmds.py
@@ -181,9 +181,15 @@ def debug_cache_stats():
     except ImportError:
         pass
 
+    tabbed_browser = objreg.get('tabbed-browser', scope='window',
+                                window='last-focused')
+    tabbed_browser_info = tabbed_browser.tabBar(). \
+        _minimum_tab_size_hint_helper.cache_info()
+
     log.misc.debug('is_valid_prefix: {}'.format(prefix_info))
     log.misc.debug('_render_stylesheet: {}'.format(render_stylesheet_info))
     log.misc.debug('history: {}'.format(history_info))
+    log.misc.debug('tab width cache: {}'.format(tabbed_browser_info))
 
 
 @cmdutils.register(debug=True)