--- a/tests/unit/mainwindow/test_tabwidget.py
+++ b/tests/unit/mainwindow/test_tabwidget.py
@@ -64,9 +64,4 @@ class TestTabWidget:
         with qtbot.waitExposed(widget):
             widget.show()
 
-        def bench():
-            for _a in range(1000):
-                # pylint: disable=protected-access
-                widget._update_tab_titles()
-                # pylint: enable=protected-access
-        benchmark(bench)
+        benchmark(widget._update_tab_titles)