I noticed that in newer versions of qb, `:tab-move` has become much slower (it's only noticable if you bind it to a key, I `:bind <alt-h> :tab-move -`.

I took a profile and (as I suspected) there was a regression which was partially introduced in https://github.com/qutebrowser/qutebrowser/pull/2847 (but not all).

I took a profile:

![2017-10-14-175302_1370x792_scrot](https://user-images.githubusercontent.com/4349709/31579653-89d079b6-b108-11e7-9c51-0ead0febf809.png)

So I decided to cache the minimumTabSizeHint function, which resulted in this profile

![2017-10-14-175342_1385x780_scrot](https://user-images.githubusercontent.com/4349709/31579655-a12cec8e-b108-11e7-81d5-93429c0c1ff7.png)

It still seems slower than before though....

This is a super low priority thing, so please handle other PR's first since they are more important :) 

One of the possible problems that I see is this won't have indicator and other padding changes take effect until text or icons change. Maybe I'll add a hook to clear the cache when those settings are reset...

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/qutebrowser/qutebrowser/3122)
<!-- Reviewable:end -->