@jgkamat 2017-10-17
I think that this isn't the only optimization we can make, since its still slow even in vertical tabs mode...

Ill keep taking a look at this, I guess, so don't merge yet :P