@The-Compiler 2017-10-20
In the latest master I added some optimizations (mainly dd927ded6ba3b03d9450c8e375b25651f67f90bf) in order to not update the tab title on every scroll move. I'm not sure whether/how this affects your profiling.

I agree the config changes should clear the cache. You can probably just adjust `_on_config_changed` to call `self.__minimumTabSizeHintHelper.cache_clear()` when needed.