@jgkamat 2017-10-21
I think that there might be a bug in the profiling software that we're using, occasionally I get those bad trees (as shown in the second image above) and that was confusing me. I tested all of these with 20 horizontal tabs (and ~5 pinned tabs among them).

I did a second profile to get better data:

current master:

![2017-10-21-001125_1884x875_scrot](https://user-images.githubusercontent.com/4349709/31848095-63e7a5a4-b5f8-11e7-9d7b-2f8fd655a1f3.png)

Shows, as before, that minimumTabSizeHint is taking up a lot of time.

After proper caching, I noticed that switching tabs is a lot faster, and startup if faster as well.

![2017-10-21-004043_1394x774_scrot](https://user-images.githubusercontent.com/4349709/31848096-7b25f6da-b5f8-11e7-8936-89dab80b9b3c.png)

I noticed there's quite a bit of duplicated work to find width of pinned tabs, so I refactored that out into one call:

![2017-10-21-004213_1389x776_scrot](https://user-images.githubusercontent.com/4349709/31848112-b0678d5e-b5f8-11e7-95aa-0b3c501ecc6f.png)

At this point, I think that the switching speed is finally acceptable with 20 tabs. I think that we can further optimize this by figuring out why the tabData is so slow, or by optimizing getting values from the completion system (I have a few experimental branches that I might clean up later with work on this).

I think that vertical tabs might be a seperate issue (and can be optimized as well).