@The-Compiler 2017-10-26
I wonder, would it be possible to write a [pytest-benchmark](http://pytest-benchmark.readthedocs.io/) test for this (search for `test_.*_benchmark` in the existing tests), so we have some numbers to look at?