@foxxx0 2017-10-26
Just did some testing with this PR and I can confirm that:

- startup with many (~20) tabs is faster
- opening and closing tabs is *way* faster
- overall the whole browser is much more responsive

So just from an end-user perspective: Thumbs up and I would really like to see this merged :)