@The-Compiler 2017-10-30
Note that you see the time for *one* invocation there. When running with `--benchmark-columns Min,Max,Median,Rounds`, here's what I see:

# No caching

## no repetition

```
------------------------ benchmark: 1 tests ------------------------
Name (time in ms)                       Min      Max  Median  Rounds
--------------------------------------------------------------------
test_update_tab_titles_benchmark     5.0693  11.3818  5.5457     181
--------------------------------------------------------------------
```

## with repetition

```
------------------------ benchmark: 1 tests -----------------------
Name (time in s)                        Min     Max  Median  Rounds
-------------------------------------------------------------------
test_update_tab_titles_benchmark     5.8613  6.3198  6.2169       5
-------------------------------------------------------------------
```

# with caching

## no repetition

```
---------------------------- benchmark: 1 tests ---------------------------
Name (time in us)                         Min         Max    Median  Rounds
---------------------------------------------------------------------------
test_update_tab_titles_benchmark     832.6320  1,743.1200  948.4280     748
---------------------------------------------------------------------------
```

## with repetition

```
--------------------------- benchmark: 1 tests --------------------------
Name (time in ms)                         Min       Max    Median  Rounds
-------------------------------------------------------------------------
test_update_tab_titles_benchmark     869.3795  998.1074  959.0883       5
-------------------------------------------------------------------------
```

# Overview

- No repetition: 5.5ms -> 948us (Δ * 1000: 4.6s)
- With repetition: 6.2s -> 959ms (Δ 5.25s)

So, seems about right :wink:

Anything else left here from your POV? Otherwise, I think this is ready to merge.